Addressing PR#17 review comments. changing activationTime data type to DevDouble.

+++ skabase/SKACapability/SKACapability/SKACapability.py
@@ -69,7 +69,7 @@ class SKACapability(SKAObsDevice):
     # ----------
 
     activationTime = attribute(
-        dtype='uint64',
+        dtype='double',
         unit="s",
         standard_unit="s",
         display_unit="s",
